Move methods for deferred value conversion out of DeferredValueSetter and into DeferredValueConverter, so they might be more easily re-used.  See issue #48.
