+++ onebusaway-gtfs-transformer/src/test/java/org/onebusaway/gtfs_transformer/impl/DeferredValueConverterTest.java
@@ -1,0 +1,106 @@
+package org.onebusaway.gtfs_transformer.impl;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.onebusaway.csv_entities.schema.BeanWrapperFactory;
+import org.onebusaway.gtfs.impl.GtfsRelationalDaoImpl;
+import org.onebusaway.gtfs.model.Agency;
+import org.onebusaway.gtfs.model.AgencyAndId;
+import org.onebusaway.gtfs.model.Route;
+import org.onebusaway.gtfs.model.ServiceCalendar;
+import org.onebusaway.gtfs.model.Stop;
+import org.onebusaway.gtfs.model.StopTime;
+import org.onebusaway.gtfs.model.Trip;
+import org.onebusaway.gtfs.model.calendar.ServiceDate;
+import org.onebusaway.gtfs.serialization.GtfsReader;
+import org.onebusaway.gtfs.services.GtfsMutableRelationalDao;
+
+/**
+ * Unit-test for {@link DeferredValueConverterTest}.
+ */
+public class DeferredValueConverterTest {
+  
+  private GtfsReader _reader = new GtfsReader();
+  private EntitySchemaCache _schemaCache = new EntitySchemaCache();
+  private GtfsMutableRelationalDao _dao = new GtfsRelationalDaoImpl();
+  
+  private DeferredValueConverter _converter;
+
+  @Before
+  public void setup() {
+    _schemaCache.addEntitySchemasFromGtfsReader(_reader);
+    _reader.setDefaultAgencyId("1");
+    _converter = new DeferredValueConverter(_reader, _schemaCache, _dao);
+  }
+
+  @Test
+  public void testString() {
+    assertEquals("Ze Stop", convert(new Stop(), "name", "Ze Stop"));
+  }
+
+  @Test
+  public void testDouble() {
+    Object value = convert(new Stop(), "lat", "47.1");
+    assertEquals(new Double(47.1), value);
+  }
+
+  @Test
+  public void testInteger() {
+    Object value = convert(new Stop(), "locationType", "1");
+    assertEquals(new Integer(1), value);
+  }
+
+  @Test
+  public void testCsvFieldMappingTime() {
+    Object value = convert(new StopTime(), "arrivalTime", "06:00:00");
+    assertEquals(6 * 60 * 60, ((Integer) value).intValue());
+  }
+
+  @Test
+  public void testCsvFieldMappingServiceDate() {
+    Object value = convert(new ServiceCalendar(), "startDate", "20130105");
+    assertEquals(new ServiceDate(2013, 1, 5), value);
+  }
+
+  @Test
+  public void testAgencAndId_DefaultAgencyId() {
+    Object value = convert(new Stop(), "id", "123");
+    assertEquals(new AgencyAndId("1", "123"), value);
+  }
+
+  @Test
+  public void testAgencAndId_ExistingAgencyId() {
+    Stop stop = new Stop();
+    stop.setId(new AgencyAndId("2", "456"));
+    Object value = convert(stop, "id", "123");
+    assertEquals(new AgencyAndId("2", "123"), value);
+  }
+
+  @Test
+  public void testEntity() {
+    Agency agency = new Agency();
+    agency.setId("1");
+    _dao.saveEntity(agency);
+    Route route = new Route();
+    Object value = convert(route, "agency", "1");
+    assertEquals(agency, value);
+  }
+
+  @Test
+  public void testEntity_AgencyAndId() {
+    Route route = new Route();
+    route.setId(new AgencyAndId("1", "10"));
+    _reader.injectEntity(route);
+    _dao.saveEntity(route);
+    Trip trip = new Trip();
+    Object value = convert(trip, "route", "10");
+    assertEquals(route, value);
+  }
+  
+  private Object convert(Object bean, String property, Object value) {
+    return _converter.convertValue(BeanWrapperFactory.wrap(bean), property,
+        value);
+  }
+}
